+++ python/lsst/ts/rotgui/main_window.py
@@ -124,9 +124,7 @@ class MainWindow(QMainWindow):
         self.setCentralWidget(container)
 
         # Disable the Qt close button
-        self.setWindowFlags(
-            Qt.Window | Qt.WindowMinimizeButtonHint | Qt.WindowMaximizeButtonHint
-        )
+        self.setWindowFlags(Qt.Window | Qt.WindowMinimizeButtonHint | Qt.WindowMaximizeButtonHint)
 
         self._add_tool_bar()
 
@@ -187,9 +185,7 @@ class MainWindow(QMainWindow):
 
         return log
 
-    def _get_log_file_name(
-        self, default_log_dir: str = "/rubin/rotator/log"
-    ) -> pathlib.Path:
+    def _get_log_file_name(self, default_log_dir: str = "/rubin/rotator/log") -> pathlib.Path:
         """Get the log file name.
 
         Parameters
@@ -206,10 +202,7 @@ class MainWindow(QMainWindow):
         log_dir = pathlib.Path(default_log_dir)
         if not log_dir.is_dir():
             print(
-                (
-                    f"Default log directory: {default_log_dir} does not exist. "
-                    "Use the home directory instead."
-                )
+                (f"Default log directory: {default_log_dir} does not exist. Use the home directory instead.")
             )
             log_dir = pathlib.Path.home()
 
@@ -274,9 +267,7 @@ class MainWindow(QMainWindow):
         action_connect.setToolTip("Connect to the rotator controller")
 
         action_disconnect = tool_bar.addAction("Disconnect", self._callback_disconnect)
-        action_disconnect.setToolTip(
-            "Disconnect and close all tasks (this might take some time)"
-        )
+        action_disconnect.setToolTip("Disconnect and close all tasks (this might take some time)")
 
         action_settings = tool_bar.addAction("Settings", self._callback_settings)
         action_settings.setToolTip("Show the application settings")
@@ -358,9 +349,7 @@ class MainWindow(QMainWindow):
         action_connect.setEnabled(False)
 
         if self.model.is_connected():
-            await prompt_dialog_warning(
-                "_callback_connect()", "The controller is already connected."
-            )
+            await prompt_dialog_warning("_callback_connect()", "The controller is already connected.")
 
         else:
             try: