+++ python/lsst/ts/rotgui/tab/tab_config.py
@@ -73,7 +73,6 @@ class TabConfig(TabTemplate):
         self._set_signal_config(self.model.signals["config"])  # type: ignore[arg-type]
 
     def create_layout(self) -> QVBoxLayout:
-
         layout = QVBoxLayout()
         layout.addWidget(self._create_group_configuration())
 
@@ -89,29 +88,19 @@ class TabConfig(TabTemplate):
         """
 
         layout = QFormLayout()
-        layout.addRow(
-            "Maximum velocity limit:", self._configuration["max_velocity_limit"]
-        )
+        layout.addRow("Maximum velocity limit:", self._configuration["max_velocity_limit"])
         layout.addRow("Velocity limit:", self._configuration["velocity_limit"])
         layout.addRow("Acceleration limit:", self._configuration["acceleration_limit"])
 
         self.add_empty_row_to_form_layout(layout)
 
-        layout.addRow(
-            "Upper position limit:", self._configuration["upper_position_limit"]
-        )
-        layout.addRow(
-            "Lower position limit:", self._configuration["lower_position_limit"]
-        )
-        layout.addRow(
-            "Disable limit maximum time:", self._configuration["disable_limit_max_time"]
-        )
+        layout.addRow("Upper position limit:", self._configuration["upper_position_limit"])
+        layout.addRow("Lower position limit:", self._configuration["lower_position_limit"])
+        layout.addRow("Disable limit maximum time:", self._configuration["disable_limit_max_time"])
 
         self.add_empty_row_to_form_layout(layout)
 
-        layout.addRow(
-            "Position error threshold:", self._configuration["position_error_threshold"]
-        )
+        layout.addRow("Position error threshold:", self._configuration["position_error_threshold"])
         layout.addRow(
             "Following error threshold:",
             self._configuration["following_error_threshold"],
@@ -120,9 +109,7 @@ class TabConfig(TabTemplate):
             "Tracking success threshold:",
             self._configuration["tracking_success_threshold"],
         )
-        layout.addRow(
-            "Tracking lost timeout:", self._configuration["tracking_lost_timeout"]
-        )
+        layout.addRow("Tracking lost timeout:", self._configuration["tracking_lost_timeout"])
 
         self.add_empty_row_to_form_layout(layout)
 
@@ -130,9 +117,7 @@ class TabConfig(TabTemplate):
             "Emergency acceleration limit:",
             self._configuration["emergency_acceleration_limit"],
         )
-        layout.addRow(
-            "Emergency jerk limit:", self._configuration["emergency_jerk_limit"]
-        )
+        layout.addRow("Emergency jerk limit:", self._configuration["emergency_jerk_limit"])
 
         self.add_empty_row_to_form_layout(layout)
 
@@ -171,43 +156,21 @@ class TabConfig(TabTemplate):
             Configuration.
         """
 
-        self._configuration["max_velocity_limit"].setText(
-            f"{config.max_velocity_limit} deg/sec"
-        )
-        self._configuration["velocity_limit"].setText(
-            f"{config.velocity_limit} deg/sec"
-        )
-        self._configuration["acceleration_limit"].setText(
-            f"{config.accel_limit} deg/sec^2"
-        )
+        self._configuration["max_velocity_limit"].setText(f"{config.max_velocity_limit} deg/sec")
+        self._configuration["velocity_limit"].setText(f"{config.velocity_limit} deg/sec")
+        self._configuration["acceleration_limit"].setText(f"{config.accel_limit} deg/sec^2")
 
-        self._configuration["position_error_threshold"].setText(
-            f"{config.pos_error_threshold} deg"
-        )
-        self._configuration["upper_position_limit"].setText(
-            f"{config.upper_pos_limit} deg"
-        )
-        self._configuration["lower_position_limit"].setText(
-            f"{config.lower_pos_limit} deg"
-        )
-        self._configuration["disable_limit_max_time"].setText(
-            f"{config.disable_limit_max_time} sec"
-        )
+        self._configuration["position_error_threshold"].setText(f"{config.pos_error_threshold} deg")
+        self._configuration["upper_position_limit"].setText(f"{config.upper_pos_limit} deg")
+        self._configuration["lower_position_limit"].setText(f"{config.lower_pos_limit} deg")
+        self._configuration["disable_limit_max_time"].setText(f"{config.disable_limit_max_time} sec")
 
-        self._configuration["following_error_threshold"].setText(
-            f"{config.following_error_threshold} deg"
-        )
+        self._configuration["following_error_threshold"].setText(f"{config.following_error_threshold} deg")
 
-        self._configuration["tracking_lost_timeout"].setText(
-            f"{config.tracking_lost_timeout} sec"
-        )
-        self._configuration["tracking_success_threshold"].setText(
-            f"{config.track_success_pos_threshold} deg"
-        )
+        self._configuration["tracking_lost_timeout"].setText(f"{config.tracking_lost_timeout} sec")
+        self._configuration["tracking_success_threshold"].setText(f"{config.track_success_pos_threshold} deg")
 
-        self._configuration["emergency_jerk_limit"].setText(
-            f"{config.emergency_jerk_limit} deg/sec^3"
-        )
+        self._configuration["emergency_jerk_limit"].setText(f"{config.emergency_jerk_limit} deg/sec^3")
         self._configuration["emergency_acceleration_limit"].setText(
             f"{config.emergency_accel_limit} deg/sec^2"
         )